1、科研成果表添加内容（INDUSTRIAL_DEVELOP_CONTENT）字段 2、调整产业发展后端代码 3、增加分组校验，不想使用分组校验的无影响 4、修复部分bug

+++ src/main/java/com/zyyglxt/service/impl/IdustrialDevelopTalRecServiceImpl.java
@@ -3,11 +3,18 @@ package com.zyyglxt.service.impl;
 import com.zyyglxt.dao.IndustrialDevelopTalRecDOMapper;
 import com.zyyglxt.dataobject.IndustrialDevelopTalRecDOKey;
 import com.zyyglxt.dataobject.IndustrialDevelopTalRecDOWithBLOBs;
+import com.zyyglxt.dataobject.validation.ValidationGroups;
+import com.zyyglxt.error.BusinessException;
+import com.zyyglxt.error.EmBusinessError;
 import com.zyyglxt.service.IIndustrialDevelopTalRecService;
+import com.zyyglxt.validator.ValidatorImpl;
+import com.zyyglxt.validator.ValidatorResult;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
 import java.util.Date;
+import java.util.List;
+import java.util.UUID;
 
 /**
  * @Author lrt
@@ -19,24 +26,43 @@ public class IdustrialDevelopTalRecServiceImpl implements IIndustrialDevelopTalR
     @Resource
     IndustrialDevelopTalRecDOMapper developTalRecDOMapper;
 
+    @Resource
+    ValidatorImpl validator;
     @Override
-    public void addTalRec(IndustrialDevelopTalRecDOWithBLOBs developTalRecDO) {
-        developTalRecDO.setCreater("未定义");
-        developTalRecDO.setUpdater("未定义");
-        developTalRecDO.setItemcreateat(new Date());
-        developTalRecDO.setItemupdateat(new Date());
-        developTalRecDOMapper.insertSelective(developTalRecDO);
+    public void addTalRec(IndustrialDevelopTalRecDOWithBLOBs record) {
+        record.setCreater("未定义");
+        record.setUpdater("未定义");
+        ValidatorResult result = validator.validate(record, ValidationGroups.Insert.class);
+        if (result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
+        if (record.getItemcode() == null || record.getItemcode().isEmpty()){
+            record.setItemcode(UUID.randomUUID().toString());
+        }
+        record.setItemcreateat(new Date());
+        record.setItemupdateat(new Date());
+
+        developTalRecDOMapper.insertSelective(record);
     }
 
     @Override
     public void delTalRec(IndustrialDevelopTalRecDOKey key) {
+        ValidatorResult result = validator.validate(key,ValidationGroups.UpdateOrDelete.class);
+        if (result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         developTalRecDOMapper.deleteByPrimaryKey(key);
     }
 
     @Override
-    public void updTalRec(IndustrialDevelopTalRecDOWithBLOBs developTalRecDO) {
-        developTalRecDO.setUpdater("未定义");
-        developTalRecDO.setItemupdateat(new Date());
-        developTalRecDOMapper.updateByPrimaryKeySelective(developTalRecDO);
+    public void updTalRec(IndustrialDevelopTalRecDOWithBLOBs record) {
+        record.setUpdater("未定义");
+        record.setItemupdateat(new Date());
+        developTalRecDOMapper.updateByPrimaryKeySelective(record);
+    }
+
+    @Override
+    public List<IndustrialDevelopTalRecDOWithBLOBs> getTalRecs() {
+        return developTalRecDOMapper.selectAll();
     }
 }
